affinex/requirements.txt
asgiref==3.8.1
crispy-bootstrap4==2024.1
Django==5.0.6
django-crispy-forms==2.3
djangorestframework==3.15.1
et-xmlfile==1.1.0
gh==0.0.4
gitdb==4.0.11
GitPython==3.1.43
joblib==1.4.2
Markdown==3.6
mistletoe==1.4.0
mysqlclient==2.2.4
numpy==2.0.1
openpyxl==3.1.5
pandas==2.2.2
pillow==10.4.0
pycodestyle==2.11.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
scikit-learn==1.5.1
scipy==1.14.0
six==1.16.0
smmap==5.0.1
sqlparse==0.5.0
threadpoolctl==3.5.0
tzdata==2024.1
